backend/requirements.txt
flask
flask-cors
python-dotenv
textblob
google-generativeai